feat: simplify api
Now we follow SRP, behaviour kept (all tests still pass).
